+++ api/authenticator.py
@@ -1,0 +1,13 @@
+
+class Authenticator():
+    def __init__(self):
+        pass
+
+    def get_token():
+        pass
+
+    def authenticate_user():
+        pass
+
+    def apply_role():
+        pass