+++ api/userManager.py
@@ -1,0 +1,21 @@
+class UserManager():
+    def __init__(self):
+        pass
+
+    def create_user(user):
+        pass
+
+    def get_user(self):
+        pass
+
+    def update_user(self):
+        pass
+
+    def delete_user(self):
+        pass
+
+    def __hash_user_password(self):
+        pass
+
+    def __generate_hash_salt(self):
+        pass